Checking Maximum of 3 Float objects using CompareTo method

+++ src/Main.java
@@ -5,12 +5,12 @@ public class Main {
         System.out.println("Find Maximum Problem using Generics");
         Scanner sc=new Scanner(System.in);
         System.out.println("Enter the first number: ");
-        Integer a= sc.nextInt();
+        Float a= sc.nextFloat();
         System.out.println("Enter the Second number: ");
-        Integer b= sc.nextInt();
+        Float b= sc.nextFloat();
         System.out.println("Enter the Third number: ");
-        Integer c= sc.nextInt();
-        Integer out=IntMax.FindMax(a,b,c);
+        Float c= sc.nextFloat();
+        Float out=IntMax.FindMax(a,b,c);
         System.out.println(out+" is the maximum number among "+a+","+b+","+c);
     }
 }
